Adding "safety valve" on streambuf notify decode payload phase
Specifically, streambuf flavor of decode-notify expects the notified
party to consume the payload from the streambuf.  If this doesn't happen,
the state machine will stay in payload mode until someone does.  If
not careful, this can infinite loop - so we empty out whatever remains
at the end of the decode-notify-payload phase.

We do this because we expect if they ignore the payload during payload
present notification, that they are never gonna be interested.